the-tool: add After=systemd-tmpfiles-setup.service
The-tool runs snap-bootstrap which needs to extract a bunch of
files to /tmp. But because there is a missing
```
After=systemd-tmpfiles-setup.service
```
the snap-bootstrap runs in parallel with
systemd-tmpfiles-setup.service which has a: `D /tmp` line
that will clean /tmp in parallel with snap-bootstrap trying
to use it.

This commit adds the missing relation.