the-tool.service: add deps on systemd-modules-load
This is not a Requires because during you might have situations where not all of
the modules loaded and hence systemd-modules-load fails, but in this case you
should still have enough to get further in the boot.

This is also what was in the service before a refactoring and was inadvertently
dropped.

Fixes https://bugs.launchpad.net/snapd/+bug/1875968.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>