many: add spread testing using GCE backend for ubuntu-core-initramfs
This sets up the project to be testable via spread, both locally via nested QEMU
VM's, as well as in the cloud with GCE nested VM's.

There are some TODO's still around simplifying this even more and getting rid of
more cruft from snapd's setup, but this is sufficient right now to get a nested
UC20 VM with our version of the initramfs in it for testing.

Still to do is to write documentation for how to run the spread tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>