factory: bootchart: start bootchart earlier
Previously, we started systemd-bootchart by adding
rd.systemd.wants=systemd-bootchart.service to the kernel command
line. This has as a side effect that the service is added to the
initrd default target and is started later than desired. Instead,
enable the service by default and use a new kernel command line
parameter (core.bootchart) as condition to start it. We also need to
make sure that the service is stopped properly to the bootchart is
written to the /run folder.