debian: make sure we get deps from compiled systemd
We were pulling some additional libraries (gcrypt and gpg-error) that
are used by Ubuntu normal systemd, but not by core-initrd's
vendordized systemd, which compiles with -Dgcrypt=false. Make sure
that we get the dependencies from the compiled systemd instead of the
ones from the build system, by setting properly LD_LIBRARY_PATH before
calling dracut-install. We need to copy the systemd libraries to a
temporal folder because dracut-install fails silently if a dependency
that is going to be pulled is inside the destination directory.