ubuntu-core-initramfs: run cpio with subprocess.run()
It has been found that in the x86 case (that prepends an early
initramfs with microcode files), the created initramfs was in some
cases not correct, and the kernel was unable to load it. Looking at
python dcoumentation about subprocess.check_call() [1], it says

"Code needing to capture stdout or stderr should use run() instead"

And it turns out that capturing output from run() fixes the problem,
although it is unclear what is the problem.