factory: add script to load drm driver from plymouth-start
Add script to load drm driver from plymouth-start service. The kernel
snap is mounted when plymouth is about to start, so we can leverage
that. We search for the boot framebuffer and load only this driver as
we do not want to delay the boot by a full run of
"udevadm trigger --type=devices --action=add", that would force
loading of all applicable kernel modules.

For this, we search the boot_vga file. This file only applies to PCI
GPUs, so in other cases we would need the drm driver to be included in
the initramfs.