Kill recovery-chooser-trigger "gracefully" so it can restart
`recovery-chooser-trigger` will be stopped when isolating to target
`initrd-switch-root.target`. The problem is that the exit status of
the process will be considered a failure. During main boot,
`recovery-chooser-trigger` will not be restarted because the stateful
reexecution will remember the failure.

The long term solution will be the make `recovery-chooser-trigger`
survive the switch root instead.