recovery-chooser-trigger.service: Survive the switch root
`snap-bootstrap recovery-chooser-trigger` is only expected to use
`/run` so it can safely survive the switch root.

When using stateful re-execution of systemd, this will keep the same
service started from initramfs and continue into the main boot.

Because we need to pull it into `initrd-switch-root.target`, we also
need to remove default dependencies so that we do not indirectly pull
`sysinit.target` or `local-fs.target` into
`initrd-switch-root.target`.

Note, we should probably add a signal handler in `snap-bootstrap
recovery-chooser-trigger` to be notified when `initrd-root-fs.target`
is reached to `chdir` into `/sysroot` so that we do not drag the old
initramfs in memory until it is stopped. This is because systemd will
lazily umount the old root.