Backport patches from sd-stub v250
Backport patches for systemd stub from systemd v250. These patches add
support to arm64 kernels by using a new kernel feature [1] for passing
the address of the initramfs to the kernel stub by a special device
path using the LoadFile2 EFI protocol. This feature is now used also
for x86. Support is in kernel 5.7 for arm64 and in 5.8 for x86.

The patches included are those affecting the src/boot/efi/ folder up
to commit [2]. Note that we are patching the Ubuntu package, which
uses as base
https://github.com/systemd/systemd-stable/releases/tag/v249.10 , so
some patches from v250 are already backported there. Additionally,
[3] from post-v250 has been included as well.

When backporting the patches, [4] has not been included because we
need to set a PCR different to the one in grub for measuring kernel
command line. An issue for this problem [5] has been opened in systemd
upstream. Also, [6] was not included either because it touched many
paths outside src/boot/efi/ and it is a cosmetic change.

Finally, some commits that do not touch src/boot/efi/ were also
included as they were needed because they do some refactoring of
macros that affect the sd-stub code. These are:

0003-alloc-util-make-mfree-typesafe.patch 5afcf89ca29b518eb2fa244b015afc2708f77e1d
0004-macro-Move-some-macros-to-macro-fundamental.h.patch f862e847246b5588d9d6ed7d91da11b6adbf39e7
0069-fundamental-define-size_t-and-memcpy-for-sd-boot.patch 5d8a725b0800ce572bb3308c03e98561251a9284
0071-move-mfree-to-macro-fundamentals.h.patch 200b1d997d96179aba9489ce9d373e869557460e

[1] https://github.com/torvalds/linux/commit/ec93fc371f014a6fb483e3556061ecad4b40735c
[2] https://github.com/systemd/systemd/commit/33bc9b756ea9a53c31340a7dcc6bc6b45c0fb2f5
[3] https://github.com/systemd/systemd/commit/178d598b5fae36fa9d54c30668771f9c626724f6
[4] https://github.com/systemd/systemd/commit/faacf1807e8fcbee4bf60495e0145e4522d76393
[5] https://github.com/systemd/systemd/issues/22635
[6] https://github.com/systemd/systemd/commit/fce9abb22793505fbb09eccd0b1e24b5d578dbcc