Move mount units into /etc/fstab
Unfortunately because /etc/fstab is still generated from
populate-writable, we need to dynamically write the missing entries.
They will be moved to core-base when we will use a static /etc/fstab.

For the moment adding the entries in /etc/fstab makes systemd be able
to keep track or where they come from (that is fstab). So we can
eventually add options so that the mount file gets re-generated with
all information into the main boot.

populate-writable gets moved before initrd-root-fs so that systemd
does parse /etc/fstab before the root switch.