Adapt maintainer scripts for ease of snap automation
As part of automating and producing more kernel snaps for more
targets, it is desired to auto-generate kernel.efi as unsigned, and
world-readable. Neither unsigned or snakeoil kernel.efi are ever
installed on end user systems, and this .deb packaging hook is only
used as part of kernel team's build processes.

Also add support for arm64, as we do produce many EFI arm64 snaps
already, which use more complicated packaging, because of this missing
create-efi call.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>