factory: create mount units specific to classic
On mantic, systemd has changed behavior and it does not run
daemon-reload from the initrd-parse-etc service if there is nothing
that specifically needs that in /etc/fstab, so we were not running
kernel-snap-generator again at this point, which meant we did not
mount /lib/{modules,firmware} on classic systems. To fix that, force
re-running the generators from a service that runs after we have the
needed mountpoints.