factory: handle properly snapd_recovery_mode based on modeenv value
The real source of true for the current mode is stored in /var/lib/snapd/modeenv
this value might be altered after
/usr/lib/systemd/system-generators/journald-console was already run.
Or /var/lib/snapd/modeenv did not even exist at that time.
This happens in install/factory-reset/recovery modes which might be
handled within the initrd.
If this is detected, update journald-console generator and remove
already created journald.conf.d/core-override.conf overlay.

Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>