extra-paths: do not mark /writable as initrd in fstab
Since populate-writable.service depends on /sysroot/writable to be
mounted, but it will create /writable entry in fstab. The one created
in fstab should not be tagged as x-initrd.mount. Otherwise it will
hide the other one and confuse systemd in potentially restarting the
mount.

This can be cleaned up once we get rid of writable-paths.