EFI stub: Relocate kernel image to allocate space for BSS
The kernel image needs to be relocated in order to allocate enough
space for the BSS. If no space is reserved for the BSS, it can overlap
with the initrd in memory, causing the corruption of it.

This change uses a similar approach as taken by upsteam in:

https://github.com/systemd/systemd/pull/33882
https://github.com/systemd/systemd/issues/33816

And fixed by systemd upstream commit id 19812661f1f6 ("stub: allocate
and zero enough space in legacy x86 handover protocol").

However since the base code was vastly different, this patch
basically replicates the same relocation strategy used by ARM in
"vendor/systemd/src/boot/efi/linux.c".

The code could have been refactored to share code with ARM, but to
minimize the risk of regressions to other architectures, linux.c was
left untouched.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>